updated properties and fixed movement warnings

+++ SFML/Player.h
@@ -4,14 +4,14 @@ class Player
 {
 public:
 	Player(int _windowWidth, int _windowHeight);
-	void spawn(int posX, int posY);
+	void spawn(float posX, float posY);
 	void draw(sf::RenderWindow &window);
 	void move(double deltaTime);
 
 private:
 	int windowWidth;
 	int windowHeight;
-	float speed = 500.0f; // when changing speed rebuild solution before compiling
+	double speed = 500; // when changing speed rebuild solution before compiling
 	sf::Sprite playerSprite;
 	sf::Texture playerTexture;
 };
